Added restart feature and saves final folders [1T']
I added a flag "int restart" at beginning of main. Setting to true will make the system not generate a random initial population (instead using a population ga.in already there). Also changed moga.c so that the last set of folders are not deleted, for debug purposes.

Also, I moved main up to the top. Makes it easier to read and easier to add a potential future header file.